Comments and cosmetic code changes.

+++ Classes/Dice.cpp
@@ -10,7 +10,7 @@ void Dice::roll() { value = roll_the_dice(); }
 // simple getter
 uint8_t Dice::get_value() const { return this->value; }
 
-void Dice::set_value(int value) { this->value = value; }
+void Dice::set_value(uint8_t value) { this->value = value; }
 
 bool Dice::get_selected() const { return this->selected; }
